requirements.txt
adbutils==1.0.9
airtest==1.2.7
allure-pytest==2.13.2
allure-python-commons==2.13.2
altgraph==0.17.3
apkutils2==1.0.0
Appium-Python-Client==2.7.1
astroid==2.15.1
async-generator==1.10
attrs==22.1.0
autopep8==1.6.0
bcrypt==4.0.1
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cigam==0.0.3
colorama==0.4.6
comtypes==1.1.14
cryptography==38.0.3
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.6
docker==4.2.2
dockerpty==0.4.1
docopt==0.6.2
docstring-to-markdown==0.12
et-xmlfile==1.1.0
exceptiongroup==1.0.1
facebook-wda==1.4.6
filelock==3.8.0
flake8==6.0.0
future==0.18.2
gitdb2==2.0.6
GitPython==2.1.15
h11==0.14.0
hrpc==1.0.9
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonpath==0.82
jsonschema==3.2.0
lazy-object-proxy==1.9.0
logzero==1.7.0
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
mss==6.1.0
nose==1.3.7
numpy==1.23.4
opencv-contrib-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.5.1
paramiko==2.12.0
parso==0.8.3
pefile==2022.5.30
Pillow==9.3.0
platformdirs==3.2.0
pluggy==1.0.0
pocoui==1.0.88
progress==1.6
py==1.11.0
pyaml==19.12.0
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.2.3
pyelftools==0.29
pyflakes==3.0.1
PyGithub==1.45
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyJWT==2.6.0
pylint==2.17.1
PyNaCl==1.5.0
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.2.1
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-ordering==0.6
pytest-rerunfailures==11.1.1
python-dateutil==2.8.2
python-gitlab==1.15.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
pytoolconfig==1.2.5
pytz==2022.6
pywin32==305
pywin32-ctypes==0.2.0
pywinauto==0.6.3
PyYAML==6.0
requests==2.28.1
retry==0.9.2
rope==1.7.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.6.0
six==1.16.0
smmap==5.0.0
smmap2==3.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
svn==1.0.1
texttable==1.6.5
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
uiautomator2==2.16.19
ujson==5.7.0
urllib3==1.26.12
websocket-client==0.48.0
whatthepatch==1.0.4
whichcraft==0.6.1
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
yapf==0.32.0
